refactor: split db/src/metadata_overrides.rs into submodules (#400)
The 893-line file is converted into a metadata_overrides/ subdirectory mirroring
the existing db/src/sync/ and db/src/discovery/ patterns:

- mod.rs    — re-exports the public API; module doc
- upsert.rs — SQL upsert/delete/query layer + apply_overrides + cover writers
- fts.rs    — FTS5 rebuild helpers (rebuild_fts_for_book extracted into
              prepare_fts_row + delete_stale_fts_row + insert_fresh_fts_row)
- links.rs  — series/author/tag link materialization (materialize_series_link
              extracted into lookup_book_id + find_or_create_series +
              link_book_to_series)
- tests.rs  — sibling test file (moved out of the inline #[cfg(test)] block)

Public API surface unchanged. The sqlx::Error boundary leak in this module is
left for the follow-up tracked in #399.

Closes #400
